// src/types/generatedAvatar.ts
export type GeneratedAvatar = {
    id: string;
    user_id: string;
    video_id: string | null;
    audio_id: string;
    image_id: string | null;
    bucket_s3: string;
    path_s3: string;
    type: string;
};

export type Question = {
    avatar_generation_id: string;
    cv_id: string;
    ground_truths: string;
    id: string;
    jd_id: string;
    question_context: string;
    topic: number;
};

// Generated by https://quicktype.io

export interface Answer {
    id: string;
    question_id: string;
    bucket_s3: string;
    video_url: string;
    audio_url: string;
    overall_score: number;
    confidence_score: number;
    text_relevancy_score: number;
    has_bad_words: boolean;
    professional_score: number;
    emotion_from_text: string;
    emotion_from_audio: string;
    emotion_from_video: string;
}

export type CVType = {
    id: string;
    user_id: string;
    full_name: string;
    email: string;
    phone_number: string;
    description: string;
    optional_infor: NullableString;
    texts: NullableString;
};

export type JDType = {
    id: string;
    title: string;
    text: string[];
};

export type NullableFile = File | null;

export type NullableBlob = Blob | null;

export type NullableString = string | null;